Schema2 SQL variants query interface evolution.
