+++ dae/dae/query_variants/sql/schema2/base_variants.py
@@ -2,18 +2,24 @@ import logging
 import abc
 import configparser
 
-from typing import Optional
+from typing import Optional, Any
 
 import pandas as pd
 
 from dae.pedigrees.family import FamiliesData
 from dae.pedigrees.loader import FamiliesLoader
+from dae.query_variants.query_runners import QueryRunner
+from dae.inmemory_storage.raw_variants import RawFamilyVariants
+from dae.query_variants.sql.schema2.family_builder import FamilyQueryBuilder
+from dae.query_variants.sql.schema2.summary_builder import SummaryQueryBuilder
 
 logger = logging.getLogger(__name__)
 
 
 class SqlSchema2Variants(abc.ABC):
     """Base class for Schema2 SQL like variants' query interface."""
+
+    RUNNER_CLASS: type[QueryRunner]
 
     def __init__(
             self,
@@ -101,3 +107,178 @@ class SqlSchema2Variants(abc.ABC):
                 ].split(",")
             ),
         }
+
+    @abc.abstractmethod
+    def _get_connection_factory(self) -> Any:
+        """Return the connection factory for specific SQL engine."""
+
+    @abc.abstractmethod
+    def _deserialize_summary_variant(self, record):
+        """Deserialize a summary variant from SQL record."""
+
+    @abc.abstractmethod
+    def _deserialize_family_variant(self, record):
+        """Deserialize a family variant from SQL record."""
+
+    # pylint: disable=too-many-arguments
+    def build_summary_variants_query_runner(
+            self,
+            regions=None,
+            genes=None,
+            effect_types=None,
+            family_ids=None,
+            person_ids=None,
+            inheritance=None,
+            roles=None,
+            sexes=None,
+            variant_type=None,
+            real_attr_filter=None,
+            ultra_rare=None,
+            frequency_filter=None,
+            return_reference=None,
+            return_unknown=None,
+            limit=None) -> QueryRunner:
+        """Build a query selecting the appropriate summary variants."""
+        query_builder = SummaryQueryBuilder(
+            self.dialect,
+            self.db,
+            self.family_variant_table,
+            self.summary_allele_table,
+            self.pedigree_table,
+            self.family_variant_schema,
+            self.summary_allele_schema,
+            self.table_properties,
+            self.pedigree_schema,
+            self.ped_df,
+            gene_models=self.gene_models,
+            do_join_affected=False,
+        )
+
+        query = query_builder.build_query(
+            regions=regions,
+            genes=genes,
+            effect_types=effect_types,
+            family_ids=family_ids,
+            person_ids=person_ids,
+            inheritance=inheritance,
+            roles=roles,
+            sexes=sexes,
+            variant_type=variant_type,
+            real_attr_filter=real_attr_filter,
+            ultra_rare=ultra_rare,
+            frequency_filter=frequency_filter,
+            return_reference=return_reference,
+            return_unknown=return_unknown,
+            limit=limit,
+        )
+        logger.debug("SUMMARY VARIANTS QUERY: %s", query)
+
+        # pylint: disable=protected-access
+        runner = self.RUNNER_CLASS(
+            self._get_connection_factory(), query,
+            deserializer=self._deserialize_summary_variant)
+
+        filter_func = RawFamilyVariants.summary_variant_filter_function(
+            regions=regions,
+            genes=genes,
+            effect_types=effect_types,
+            family_ids=family_ids,
+            person_ids=person_ids,
+            inheritance=inheritance,
+            roles=roles,
+            sexes=sexes,
+            variant_type=variant_type,
+            real_attr_filter=real_attr_filter,
+            ultra_rare=ultra_rare,
+            frequency_filter=frequency_filter,
+            return_reference=return_reference,
+            return_unknown=return_unknown,
+            limit=limit)
+
+        runner.adapt(filter_func)
+
+        return runner
+
+    # pylint: disable=too-many-arguments,too-many-locals
+    def build_family_variants_query_runner(
+            self,
+            regions=None,
+            genes=None,
+            effect_types=None,
+            family_ids=None,
+            person_ids=None,
+            inheritance=None,
+            roles=None,
+            sexes=None,
+            variant_type=None,
+            real_attr_filter=None,
+            ultra_rare=None,
+            frequency_filter=None,
+            return_reference=None,
+            return_unknown=None,
+            limit=None,
+            pedigree_fields=None):
+        """Build a query selecting the appropriate family variants."""
+        do_join_pedigree = pedigree_fields is not None
+        query_builder = FamilyQueryBuilder(
+            self.dialect,
+            self.db,
+            self.family_variant_table,
+            self.summary_allele_table,
+            self.pedigree_table,
+            self.family_variant_schema,
+            self.summary_allele_schema,
+            self.table_properties,
+            self.pedigree_schema,
+            self.ped_df,
+            gene_models=self.gene_models,
+            do_join_pedigree=do_join_pedigree,
+        )
+
+        query = query_builder.build_query(
+            regions=regions,
+            genes=genes,
+            effect_types=effect_types,
+            family_ids=family_ids,
+            person_ids=person_ids,
+            inheritance=inheritance,
+            roles=roles,
+            sexes=sexes,
+            variant_type=variant_type,
+            real_attr_filter=real_attr_filter,
+            ultra_rare=ultra_rare,
+            frequency_filter=frequency_filter,
+            return_reference=return_reference,
+            return_unknown=return_unknown,
+            limit=limit,
+            pedigree_fields=pedigree_fields
+        )
+
+        logger.debug("FAMILY VARIANTS QUERY: %s", query)
+        deserialize_row = self._deserialize_family_variant
+
+        # pylint: disable=protected-access
+        runner = self.RUNNER_CLASS(
+            self._get_connection_factory(), query,
+            deserializer=deserialize_row)
+
+        filter_func = RawFamilyVariants.family_variant_filter_function(
+            regions=regions,
+            genes=genes,
+            effect_types=effect_types,
+            family_ids=family_ids,
+            person_ids=person_ids,
+            inheritance=inheritance,
+            roles=roles,
+            sexes=sexes,
+            variant_type=variant_type,
+            real_attr_filter=real_attr_filter,
+            ultra_rare=ultra_rare,
+            frequency_filter=frequency_filter,
+            return_reference=return_reference,
+            return_unknown=return_unknown,
+            limit=limit)
+
+        runner.adapt(filter_func)
+
+        return runner
